fix: P3 audit fixes — 35 of 39 findings (docs, tracing, errors, tests, API)
P3 fixes applied:
- DOC-39/40/41/43, AD-45/46, RM-36: 10 stale E5-base/768 doc comments
- DOC-45: CONTRIBUTING.md add cqs-verify skill
- DOC-46: CQS_ONNX_DIR documented in README + SECURITY.md
- AD-50: VectorIndex::dim() trait method added
- AD-53: ModelInfo.dimensions u32→usize (removes casts everywhere)
- AD-54: Embedding::new vs try_new doc guidance
- OB-28..32: tracing spans for detect_provider, create_session,
  parse_unified_diff, find_changed_functions, audit state, embeddings batch
- EH-41/42/44: warn on silently swallowed DB/HTTP errors
- CQ-35: dead check_model_version → #[cfg(test)]
- RB-30: fragile checkpoint_sha unwrap → is_some_and
- AC-29/30: documented edge case behavior
- DS-36: HNSW backup rename failure now logged
- PERF-41: use make_placeholders helper
- RM-35: clear_session clears query cache
- EX-36: doc_format validation test for all 51 languages
- EX-37/38: download size comment, CAGRA_THRESHOLD env var override
- PB-35: onnx_path forward-slash contract documented
- SEC-26: api_base credential redaction in logs
- SEC-27: git_show path colon validation
- TC-42/44/49: 9 new tests (phantom chunks, SearchFilter, validate_f32)

Also: CLAUDE.md workflow examples, ROADMAP agent adoption updates,
pre-edit hook, v9-200k-1.5ep results, research log update.

Skipped: AD-48 (low benefit), PERF-42 (not stale), RB-32 (trait change),
EX-36 full enum (test approach instead).

+++ src/store/mod.rs
@@ -92,17 +92,17 @@ pub use helpers::UnifiedResult;
 /// Current database schema version.
 pub use helpers::CURRENT_SCHEMA_VERSION;
 
-/// Name of the embedding model (compile-time default for E5-base-v2).
+/// Name of the embedding model (compile-time default for BGE-large).
 ///
 /// Runtime code should use `Store::stored_model_name()` or `ModelInfo::new()`.
 /// This constant exists for callers outside the store (e.g. `doctor.rs`).
 pub const MODEL_NAME: &str = crate::embedder::DEFAULT_MODEL_REPO;
 
-/// Expected embedding dimensions (compile-time default for E5-base-v2).
+/// Expected embedding dimensions (compile-time default for BGE-large).
 ///
 /// Runtime code should use `Store::dim` instead. This constant exists for
 /// callers outside the store that need a compile-time value.
-pub const EXPECTED_DIMENSIONS: u32 = crate::EMBEDDING_DIM as u32;
+pub const EXPECTED_DIMENSIONS: usize = crate::EMBEDDING_DIM;
 
 /// Default name_boost weight for CLI search commands.
 pub use helpers::DEFAULT_NAME_BOOST;
